consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = "Serviço que consome dados do Kafka e armazena em banco de dados"
authors = ["Ana Bê <contato.anabeatrizoliver@gmail.com"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.2.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"